Libraries/PyKotorGL/pyproject.toml
[build-system]
requires = ["setuptools>=67.8.0,<70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pykotorgl"
version = "1.8.0"
description = "OpenGL rendering module for PyKotor - renders KotOR modules and scenes."
authors = [{name = "NickHugi"}]
maintainers = [{name = "th3w1zard1", email = "halomastar@gmail.com"}]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "LGPL-3.0-or-later"}
keywords = [
    "kotor", "kotor1", "kotor2", "tsl",
    "opengl", "rendering", "3d", "graphics",
    "mdl", "mdx", "model", "scene",
    "pykotor", "bioware",
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pykotor>=1.8.0",
    "numpy>=1.22",
    "PyOpenGL~=3.1",
    "PyGLM>=2.0,<2.8; python_implementation == 'CPython'",
]

[project.optional-dependencies]
gl_accel = ["PyOpenGL-accelerate>=3.1.10"]
moderngl = ["moderngl>=5.8.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/th3w1zard1/PyKotor"
Documentation = "https://github.com/th3w1zard1/PyKotor/wiki"
Repository = "https://github.com/th3w1zard1/PyKotor.git"
Issues = "https://github.com/th3w1zard1/PyKotor/issues"
Changelog = "https://github.com/th3w1zard1/PyKotor/releases"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

# Poetry configuration
[tool.poetry]
name = "pykotorgl"
version = "1.8.0"
description = "OpenGL rendering extension for PyKotor"
authors = ["NickHugi", "Benjamin Auquite <halomastar@gmail.com>"]
maintainers = ["Benjamin Auquite <halomastar@gmail.com>"]
readme = "README.md"
license = "LGPL-3.0-or-later"
packages = [{ include = "pykotor", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pykotor = "^1.8.0"
numpy = ">=1.22"
PyOpenGL = "~3.1"
PyGLM = {version = ">=2.0,<2.8", python = "^3.8", markers = "python_implementation == 'CPython'"}

[tool.poetry.extras]
gl_accel = ["PyOpenGL-accelerate"]
moderngl = ["moderngl"]

[tool.poetry.urls]
Homepage = "https://github.com/th3w1zard1/PyKotor"
Documentation = "https://github.com/th3w1zard1/PyKotor/wiki"
Repository = "https://github.com/th3w1zard1/PyKotor.git"
Issues = "https://github.com/th3w1zard1/PyKotor/issues"

########################################################
## define only linter configurations below this point ##
########################################################

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ruff.lint]
ignore = [
    "N",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ALL"]
"__init__.py" = ["I001", "F401"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.mccabe]
max-complexity = 25

[tool.style]
based_on_style = "pep8"
split_before_named_assigns = true
split_complex_comprehension = true
split_arguments_when_comma_terminated = true

[tool.pylintrc]
max-line-length = 175
